Run Postgres in Docker tests through shell.
Uh, this is quite a bizarre problem. Whenever I gdb attach to any process inside
a container, someone sends to another someone (and not the process being
debugged) SIGKILL. I am not sure who are those guys, but the latter is evidently
was *not* born by postmaster; however, it is a daemonized process, i.e. its
parent at that moment not its father but process with PID 1 -- postmaster, in
our case. And you know what happens when postmaster notices that its child had
died with SIGKILL, right?

Sending those SIGCHLDs to the shell instead of postmaster helps.