[PGPRO-3220] Declare the current_gtid variable as volatile.
It is needed in accordance with rules, defined in comment section of elog.h:

"Note: if a local variable of the function containing PG_TRY is modified
in the PG_TRY section and used in the PG_CATCH section, that variable
must be declared "volatile" for POSIX compliance.  This is not mere
pedantry; we have seen bugs from compilers improperly optimizing code
away when such a variable was not marked.  Beware that gcc's -Wclobbered
warnings are just about entirely useless for catching such oversights."