dmq: add sender connection counters.
Counters allow to notice send connection reset which could possibly drop request
-- and thus we should stop waiting for response. However, they don't save from
potential deadlocks -- if a loop of nodes fills all queues in one direction with
requests (reverse direction must constantly blink for that), nobody will be able
to send replies. So eventually we should have 'connection lost in one direction,
close it in the other' logic. But such deadlocks seem to be hardly possible
in practice, so not today.

gather() is adapted to new API, but nodoby uses it yet. It doesn't rely on
whether node itself and others are enabled/disabled anymore.

This also adds separate DMQ_N_MASK_POS macro, as we can actually have small
number of DMQ_MAX_DESTINATIONS with large mask positions.