Restore the referee mode, yet another time.
It was obviously broken by generations & ACP stuff.

Ideologically it is like the previous but the implementation is heavily
revised. The code is amply documented though surely there must be some
high-level design text which I will definitely add at some point.

I'll put here a list of dubious places (they are not necessarily bugs, but...)
of the previous implementation:
 - surely it worked only with 2 nodes, but this wasn't enforced anywhere;
 - that insert on conflict do nothing with comment 'overwrite' is also very
   weird;
 - return value of MtmRefereeGetWinner is never compared with my node id, oh;
 - Concurrent resolving by referee and normally? there was no gtx;
 - Forgetting to clear the grant on network failure after acquire;
 - OTOH, dangerous possibility of clearing of too new flag: we don't have
   generations, so there are no numbers to compare;
 - I see that 'caching' is for reboot of the granted node, but it looks
   somewhat suspicious;

This thing works only with 2 nodes due to how xact resolution is done; now this
is checked on monitor start. Also, referee support was added to Cluster.pm and
make start/stop (as well as configurable number of nodes BTW). Use it like

RUN_OPTS='--referee' make start