Fix rare bgwpool-related segfault.
It is quite weird to store bgwhandles pointer to private memory in shmem, but
anyway.

Independently, add logging of skipped AP due to missing xact; it shows more
clearly the race resulting in xact never being marked orphaned while it should
-- when e.g. client suddently disconnects, P might reach peer later than AP and
hang as not orphaned (while receiver is healthy, of course).