Revive add_node/drop_node.
It is just an adaptation of generations to the previous implementation. So
 - there is still hard bounding to the donor to get proper syncpoints
 - there is no protection against double membership change (no proper
   two-phase membership switch procedure)
 - it seems that other nodes might not hold enough WAL for new node if there
   is concurrent load and faults.
 - there is no protection against someone not having all xacts of dropped
   node before drop.

Due to this I've removed concurrent load from the test in the default
configuration.

However, it should be still useful for basic dead node replacements.

Also fix several adjacent bugs:
 - set configured mask in MtmConsiderGenSwitch;
 - use proper total num of nodes instead of max node id for
   mtm.max_workers handling in check_config;
 - typo in MtmAllApplyWorkersFinished;
 - send 0 instead of -1 when origin node is not known; -1 was broken as
   unsigned ints were used.

Also use 'configured' mask of the xact generation to determine quorum in
resolver.