[PGPRO-3182] Don't SetCatchupDonor when node sees the majority but not the clique.
Comment there stated this is needed for ousider node catching up in stable
sausage topology, but it is not: such outsider still sees the clique, though it
is not the clique the middle node agrees with. Moreover, since 84d6b3ff2f10 it
triggered assertion failure in SetCatchupDonor as there is a short moment after
node marked as connected but before the first heartbeat with last_online_in
comes.

tags: multimaster