[PGPRO-4819] Fix "Prevent one of AB going into recovery when C gets down".
(i.e. 40355654d205f)

The idea of refusing the vote if responder gen is >= than proposed gen num is
ok, but we didn't hint to increase the number on the next campaign, so there was
a small risk of node C infinitely proposing 9 <ABC> while both A and B live in 9
<AB> and refuse to vote. Fix by 1) making the hint (send last_vote_num); 2) to
avoid devaluing the whole idea, force learning of reponder's generation in this
case.